Move integration algorithms to integrals subfolder
