+++ desc/compute/_field.py
@@ -13,17 +13,14 @@ from scipy.constants import mu_0
 
 from desc.backend import jnp
 
-from .data_index import register_compute_fun
-from .utils import (
-    cross,
-    dot,
-    safediv,
-    safenorm,
+from ..integrals.surface_integral import (
     surface_averages,
     surface_integrals_map,
     surface_max,
     surface_min,
 )
+from .data_index import register_compute_fun
+from .utils import cross, dot, safediv, safenorm
 
 
 @register_compute_fun(
