+++ desc/compute/_geometry.py
@@ -11,8 +11,9 @@ expensive computations.
 
 from desc.backend import jnp
 
+from ..integrals.surface_integral import line_integrals, surface_integrals
 from .data_index import register_compute_fun
-from .utils import cross, dot, line_integrals, safenorm, surface_integrals
+from .utils import cross, dot, safenorm
 
 
 @register_compute_fun(
