+++ desc/integrals/singularities.py
