+++ desc/objectives/_free_boundary.py
@@ -9,13 +9,13 @@ from desc.backend import jnp
 from desc.compute import get_params, get_profiles, get_transforms
 from desc.compute.utils import _compute as compute_fun
 from desc.grid import LinearGrid
-from desc.nestor import Nestor
-from desc.objectives.objective_funs import _Objective
-from desc.singularities import (
+from desc.integrals.singularities import (
     DFTInterpolator,
     FFTInterpolator,
     virtual_casing_biot_savart,
 )
+from desc.nestor import Nestor
+from desc.objectives.objective_funs import _Objective
 from desc.utils import Timer, errorif, warnif
 
 from .normalization import compute_scaling_factors
