+++ desc/plotting.py
@@ -18,9 +18,10 @@ from desc.backend import sign
 from desc.basis import fourier, zernike_radial_poly
 from desc.coils import CoilSet, _Coil
 from desc.compute import data_index, get_transforms
-from desc.compute.utils import _parse_parameterization, surface_averages_map
+from desc.compute.utils import _parse_parameterization
 from desc.equilibrium.coords import map_coordinates
 from desc.grid import Grid, LinearGrid
+from desc.integrals.surface_integral import surface_averages_map
 from desc.magnetic_fields import field_line_integrate
 from desc.utils import errorif, only1, parse_argname_change, setdefault
 from desc.vmec_utils import ptolemy_linear_transform
