+++ tests/test_plotting.py
@@ -15,10 +15,10 @@ from desc.basis import (
 )
 from desc.coils import CoilSet, FourierXYZCoil, MixedCoilSet
 from desc.compute import data_index
-from desc.compute.utils import surface_averages
 from desc.examples import get
 from desc.geometry import FourierRZToroidalSurface, FourierXYZCurve
 from desc.grid import ConcentricGrid, Grid, LinearGrid, QuadratureGrid
+from desc.integrals.surface_integral import surface_averages
 from desc.io import load
 from desc.magnetic_fields import (
     OmnigenousField,
